requirements.txt
flask
requests
streamlit
python-dotenv
huggingface_hub
